+++ CHANGES.rst
@@ -1,9 +1,17 @@
 
-0.0.14.5 (in progress)
+0.0.14.6 (in progress)
+---------------------------
+Since last release
+   - pipline script in main repo (rc1)
+
+
+
+0.0.14.5
 ---------------------------
 Since last release
    - packaging improvements and fixing bugs introduced in restructuring (rc1-rc4)
    - fix bug in mcsamplerEnsemble (used with --sampler-method GMM) (rc5)
+   - helper can parse LI ini files 
 
 0.0.14.4 (2019-10-3)
 ------------------------------